Validacao de e-mail duplicado

+++ src/main/kotlin/com/jjcdutra/mercadolivro/infrastructure/adapters/repository/SpringAutorRepository.kt
@@ -3,7 +3,7 @@ package com.jjcdutra.mercadolivro.infrastructure.adapters.repository
 import com.jjcdutra.mercadolivro.domain.Autor
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface SpringAutorRepository : JpaRepository<Autor, Long> {
+interface SpringAutorRepository : JpaRepository<Autor, Int> {
 
     fun existsByEmail(value: String): Boolean
 }
